00 speech scraper/requirements.txt
requests
pandas
python-dateutil
ftfy
